refactor: Organize the package structure for rest
