+++ src/main/java/ryzendee/app/rest/impl/base/DealContractorToRoleRestController.java
@@ -1,10 +1,10 @@
-package ryzendee.app.rest;
+package ryzendee.app.rest.impl.base;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RestController;
 import ryzendee.app.dto.DealContractorRoleAddRequest;
 import ryzendee.app.dto.DealContractorRoleRemoveRequest;
-import ryzendee.app.rest.api.DealContractorToRoleApi;
+import ryzendee.app.rest.api.base.DealContractorToRoleApi;
 import ryzendee.app.service.DealContractorToRoleService;
 
 @RestController
