+++ src/test/java/ryzendee/app/rest/base/DealContractorRestControllerTest.java
@@ -1,4 +1,4 @@
-package ryzendee.app.rest;
+package ryzendee.app.rest.base;
 
 import io.restassured.http.ContentType;
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
@@ -6,12 +6,14 @@ import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
 import ryzendee.app.dto.ContractorSaveRequest;
 import ryzendee.app.exception.ResourceNotFoundException;
+import ryzendee.app.rest.impl.base.DealContractorRestController;
 import ryzendee.app.service.DealContractorService;
 
 import java.util.UUID;
@@ -19,6 +21,7 @@ import java.util.UUID;
 import static org.mockito.Mockito.*;
 import static ryzendee.app.testutils.FixtureUtil.contractorSaveRequestBuilderFixture;
 
+@AutoConfigureMockMvc(addFilters = false)
 @WebMvcTest(DealContractorRestController.class)
 public class DealContractorRestControllerTest {
 
@@ -79,5 +82,4 @@ public class DealContractorRestControllerTest {
 
         verify(dealContractorService).deleteById(contractorId);
     }
-
 }
