+++ src/test/java/ryzendee/app/rest/base/DealContractorToRoleRestControllerTest.java
@@ -1,4 +1,4 @@
-package ryzendee.app.rest;
+package ryzendee.app.rest.base;
 
 import io.restassured.http.ContentType;
 import io.restassured.module.mockmvc.RestAssuredMockMvc;
@@ -6,6 +6,7 @@ import io.restassured.module.mockmvc.specification.MockMvcRequestSpecification;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.HttpStatus;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -13,11 +14,13 @@ import org.springframework.test.web.servlet.MockMvc;
 import ryzendee.app.dto.DealContractorRoleAddRequest;
 import ryzendee.app.dto.DealContractorRoleRemoveRequest;
 import ryzendee.app.exception.ResourceNotFoundException;
+import ryzendee.app.rest.impl.base.DealContractorToRoleRestController;
 import ryzendee.app.service.DealContractorToRoleService;
 
 import static org.mockito.Mockito.*;
 import static ryzendee.app.testutils.FixtureUtil.*;
 
+@AutoConfigureMockMvc(addFilters = false)
 @WebMvcTest(DealContractorToRoleRestController.class)
 public class DealContractorToRoleRestControllerTest {
